requirements.txt
pyTelegramBotAPI==4.12.0
yt-dlp==2026.2.28
python-dotenv==1.0.0